refactor: generic drawer wrapper
