+++ packages/desktop/components/drawers/index.ts
@@ -1,1 +1,2 @@
+export { default as Drawer } from './Drawer.svelte'
 export { default as DrawerTemplate } from './DrawerTemplate.svelte'
