+++ packages/desktop/lib/auxiliary/drawer/actions/index.ts
@@ -1,2 +1,3 @@
 export * from './closeDrawer'
 export * from './openDrawer'
+export * from './toggleDashboardDrawer'
