+++ packages/desktop/lib/auxiliary/drawer/actions/openDrawer.ts
@@ -1,13 +1,14 @@
 import { DrawerDirection } from '../enums'
-import { IDrawerState } from '../interfaces'
 import { drawerState } from '../stores'
+import { DrawerState } from '../types'
 
 export function openDrawer({
+    route,
     id,
     direction = DrawerDirection.Right,
     hideClose = false,
     preventClose = false,
     overflow = false,
-}: Omit<IDrawerState, 'active'>): void {
-    drawerState.set({ active: true, id, hideClose, preventClose, direction, overflow })
+}: Omit<DrawerState, 'active'>): void {
+    drawerState.set({ active: true, route, id, hideClose, preventClose, direction, overflow })
 }
