+++ packages/desktop/lib/auxiliary/drawer/enums/drawer-route.enum.ts
@@ -1,0 +1,3 @@
+export enum DrawerRoute {
+    Dashboard = 'dashboard',
+}
