+++ packages/desktop/lib/auxiliary/drawer/enums/index.ts
@@ -1,1 +1,2 @@
 export * from './drawer-direction.enum'
+export * from './drawer-route.enum'
