+++ packages/desktop/lib/auxiliary/drawer/interfaces/drawer-state.interface.ts
@@ -1,11 +1,15 @@
 import { DashboardDrawerRoute } from '@views/dashboard/drawers'
-import { DrawerDirection } from '../enums'
+import { DrawerDirection, DrawerRoute } from '../enums'
 
 export interface IDrawerState {
     active: boolean
-    id: DashboardDrawerRoute
     hideClose?: boolean
     preventClose?: boolean
     direction?: DrawerDirection
     overflow?: boolean
+}
+
+export interface IDashboardDrawerState extends IDrawerState {
+    route: DrawerRoute.Dashboard
+    id: DashboardDrawerRoute
 }
