+++ packages/desktop/lib/auxiliary/drawer/types/drawer-state.type.ts
@@ -1,0 +1,3 @@
+import { IDashboardDrawerState } from '../interfaces'
+
+export type DrawerState = IDashboardDrawerState
