+++ packages/desktop/lib/auxiliary/drawer/types/index.ts
@@ -1,0 +1,1 @@
+export * from './drawer-state.type'
